Completely refactor extension to multiple files
