+++ typescript/packages/clipper-extension/src/pages/Popup.css
@@ -6,28 +6,110 @@ div {
     align-items: center;
     justify-content: center;
 }
-  
+
 img {
     width: 200px;
     height: 200px;
 }
-  
+
 h1 {
     font-size: 18px;
     color: white;
     font-weight: bold;
     margin: 0;
 }
-  
+
 p {
     color: white;
     opacity: 0.7;
     margin: 0;
 }
-  
+
 code {
     font-size: 12px;
     padding: 2px 4px;
     background-color: #ffffff24;
     border-radius: 2px;
+}
+
+.clipper-popup {
+    width: 100%;
+    height: 100%;
+}
+
+.preview-section {
+    border: 1px solid #ddd;
+    border-radius: 8px;
+    padding: 16px;
+    margin-bottom: 16px;
+    background: #f9f9f9;
+}
+
+.preview-content {
+    max-width: 100%;
+    overflow-wrap: break-word;
+    word-wrap: break-word;
+    word-break: break-word;
+}
+
+.preview-content h3 {
+    margin: 0 0 8px 0;
+    font-size: 16px;
+    color: #333;
+}
+
+.meta {
+    margin-top: 8px;
+    font-size: 12px;
+    color: #666;
+}
+
+.clipping-controls {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+}
+
+.preview-tabs {
+    display: flex;
+    gap: 8px;
+    margin-bottom: 16px;
+}
+
+.preview-tabs button {
+    padding: 8px 16px;
+    border: 1px solid #ddd;
+    background: white;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+.preview-tabs button.active {
+    background: #e0e0e0;
+}
+
+.raw-section {
+    margin-bottom: 16px;
+}
+
+.raw-section h4 {
+    margin: 0 0 8px 0;
+}
+
+.raw-section pre {
+    max-width: 320px;
+    background: #f0f0f0;
+    padding: 12px;
+    border-radius: 4px;
+    max-height: 200px;
+    overflow: auto;
+    margin: 0;
+    font-size: 12px;
+}
+
+.strategy-picker {
+    position: absolute;
+    top: 16px;
+    left: 16px;
+    z-index: 1;
 }
